apps/rag/pyproject.toml
[project]
name = "nexusai-rag"
version = "0.1.0"
description = "NexusAI hybrid RAG service (pgvector + BM25 + rerank)"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.6",
    "pydantic>=2.9.2",
    "pydantic-settings>=2.5.2",
    "httpx>=0.27.2",
    "asyncpg>=0.29.0",
    "psycopg[binary]>=3.2.3",
    "sqlalchemy[asyncio]>=2.0.35",
    "pgvector>=0.3.4",
    "tiktoken>=0.8.0",
    "google-generativeai>=0.8.2",
    "prometheus-client>=0.21.0",
    "structlog>=24.4.0",
    "numpy>=2.1.2",
    "rank-bm25>=0.2.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.3",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.6.9",
    "mypy>=1.11.2",
]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
# Keep style linting soft for a young codebase — only flag real problems.
select = ["E", "F", "I", "UP"]
ignore = [
    "E501",   # line too long — formatter's job
    "E402",   # module level import not at top of file
    "F401",   # unused import — common during rapid iteration
    "F403",   # star imports
    "F841",   # unused local variable
    "UP007",  # Union X | Y rewrites — optional
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401", "F403"]
